Upgraded dependencies, moved everything except blog into one folder
