+++ site/package.json
@@ -10,28 +10,28 @@
     "check": "tsc"
   },
   "dependencies": {
-    "framer-motion": "^6.5.1",
-    "next": "^12.2.3",
-    "next-themes": "^0.2.0",
+    "framer-motion": "^7.6.7",
+    "next": "^13.0.4",
+    "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.4.0"
+    "react-icons": "^4.6.0"
   },
   "devDependencies": {
-    "@types/node": "17.0.35",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.39.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.39.0",
-    "@typescript-eslint/parser": "^5.39.0",
-    "autoprefixer": "^10.4.8",
-    "cssnano": "^5.1.12",
-    "eslint": "^8.20.0",
-    "eslint-config-next": "^12.2.3",
-    "postcss": "^8.4.14",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "autoprefixer": "^10.4.13",
+    "cssnano": "^5.1.14",
+    "eslint": "^8.27.0",
+    "eslint-config-next": "^13.0.4",
+    "postcss": "^8.4.19",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.1.7",
-    "typescript": "^4.8.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.9.3"
   }
 }
