+++ site/pages/todo/index.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import { BsCheck2 } from "react-icons/bs";
 
-import ThemeSwitch from "./components/themeswitch";
+import ThemeSwitch from "../components/themeswitch";
 
 const TODO = [
   {
