--- site_projects/styles/globals.css
@@ -1,27 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Turret+Road:wght@200&display=swap");
-
-* {
-  transition: 150ms ease-in-out;
-}
-
-html,
-body {
-  background: #212627;
-  color: #b8c0d3;
-}
-
-[data-theme="light"],
-[data-theme="light"] body {
-  background: #c7cde4;
-  color: #3b4345;
-}
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    font-family: "Turret Road", "Montserrat";
-  }
-}
