--- site_start/.eslintrc.json
@@ -1,12 +1,0 @@
-{
-  "extends": [
-    "next/core-web-vitals",
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
-  "parserOptions": {
-    "project": "**/tsconfig.json"
-  }
-}
