--- site_start/styles/globals.css
@@ -1,32 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@300&display=swap");
-
-* {
-  transition: 150ms ease-in-out;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-html,
-body {
-  background: #212627;
-  color: #b8c0d3;
-}
-
-[data-theme="light"],
-[data-theme="light"] body {
-  background: #c7cde4;
-  color: #3b4345;
-}
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    font-family: "Montserrat", "Turret Road";
-  }
-}
