--- site_todo/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "printWidth": 100,
-  "proseWrap": "never"
-}
