--- site_todo/next.config.js
@@ -1,5 +1,0 @@
-/** @type {import('next').NextConfig} */
-module.exports = {
-  reactStrictMode: true,
-  swcMinify: true,
-};
