--- site_todo/styles/globals.css
@@ -1,37 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@300;600&display=swap");
-
-* {
-  transition: 150ms ease-in-out;
-}
-
-html,
-body {
-  background: #212627;
-  color: #b8c0d3;
-}
-
-[data-theme="light"],
-[data-theme="light"] body {
-  background: #c7cde4;
-  color: #3b4345;
-}
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    font-family: "Montserrat", "Turret Road";
-  }
-}
-
-@layer utilities {
-  [data-theme="light"] tr.stripe:nth-of-type(2n) {
-    background-color: #b9bfd4;
-  }
-
-  tr.stripe:nth-of-type(2n) {
-    background-color: #24292a;
-  }
-}
