--- site_todo/tailwind.config.js
@@ -1,7 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./pages/**/*.{js,ts,jsx,tsx}", "./components/**/*.{js,ts,jsx,tsx}"],
-  darkMode: "class",
-  theme: {},
-  plugins: [],
-};
